# server/api/openapi-with-docs.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Dead Pigeons API",
    "description": "API for the Dead Pigeons bingo project.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Players": {
      "get": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_GetAll",
        "parameters": [
          {
            "name": "onlyActive",
            "in": "query",
            "schema": {
              "type": "boolean",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_Create",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerCreateDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Players/{id}/status": {
      "patch": {
        "tags": [
          "Players"
        ],
        "operationId": "Players_UpdateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PlayerResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "playerId": {
            "type": "string",
            "format": "guid"
          },
          "fullName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "PlayerCreateDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "fullName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "nullable": true
          }
        }
      }
    }
  }
}